requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
opencv-python==4.8.1.78
matplotlib==3.7.2
scikit-image==0.21.0
pyyaml==6.0
tqdm==4.66.0
pillow==10.0.0
tensorboard==2.14.0
albumentations==1.3.0
scipy==1.11.3
